Started switching to new match execution model
In the new model, matches yield a sequence of possible "solutions", i.e. possible contexts after the execution of the match, where the text has been consumed and placed in 'matched_text', the 'position' has been advanced, etc.

Note that this completely breaks the matches codebase. In the following diffs, the matches will be converted to the new system one by one.