Always use the extension for directories
This changes the rules for directory renaming and name verification such that the extension is always included in their name regardless of the "use extension" setting. This allows for a more logical handling of directory names like "01. Title", given that, unlike files, directories generally need no extension to indicate their type, so a period in their name is usually just a title separator or has some other grammatical meaning. This also means that the handling of directories will now be completely uniform with respect to their position in the path (path component vs. leaf).