Removed backslash escaping from string literals
I've decided to switch to a much simpler Excel-style (i.e. doubling) escaping mechanism for BAON literals. The reason is that the backslash is heavily used on Windows as a path separator, standard escapes produce characters that are not allowed in filenames, and Unicode characters can just be typed directly.