Started major "matches with actions" refactoring
The point is to re-implement the "with actions" nature of matches so that, instead of actions being a property of a class of matches (i.e. implemented by inheritance), they are instead attached through a dedicated composite match (i.e. implementation by composition).

This first diff adjusts the parser and the parsing tests. The match classes still subclass MatchWithActions, but the actions field is no longer used because all actions are now captured by the eponymous composite match.